% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/log_event.R
\name{as_log_event}
\alias{as_log_event}
\alias{as_log_event.default}
\alias{as_log_event.log_event}
\alias{as_log_event.simpleError}
\alias{as_log_event.simpleMessage}
\alias{as_log_event.simpleWarning}
\title{Log event conversion methods.}
\usage{
as_log_event(event, ...)

\method{as_log_event}{simpleError}(event, ...)

\method{as_log_event}{simpleWarning}(event, ...)

\method{as_log_event}{simpleMessage}(event, ...)

\method{as_log_event}{log_event}(event, ...)

\method{as_log_event}{default}(event, ...)
}
\arguments{
\item{event}{a \code{condition} to be coersed to a \code{log_event}.}

\item{...}{Currently not used.}
}
\value{
a \code{log_event} condition object.
}
\description{
Methods for coersing conditions to log events. Methods are predefined
for \code{simpleMessage}, \code{simpleWarning}, \code{simpleError}
classes, but others can be defined. The default method will raise
an error.
}

