% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRawFile.r
\name{CleanRawFile}
\alias{CleanRawFile}
\title{Clean raw file from a logging device}
\usage{
CleanRawFile(file, outfile)
}
\arguments{
\item{file}{character Path to the raw data file}

\item{outfile}{character Path to the cleaned file (incl name of file)}
}
\description{
General clean up and subset of raw file from logger. Currently only
implemented for GiPSy-5 micro GPS dataloggers.
}

