% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{close_log}
\alias{close_log}
\title{Close a log file for loggr}
\usage{
close_log()
}
\description{
This closes a log file for \code{\link{loggr}}.
}
\examples{

open_log("mylog.txt")
close_log()

}
\seealso{
\code{\link{open_log}}
}

