% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deactivate_log.R
\name{deactivate_log}
\alias{deactivate_log}
\title{Deactivate Active Log Objects}
\usage{
deactivate_log(names = loggr_list())
}
\arguments{
\item{names}{character: vector of names of log objects to deactivate.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Deactivate Active Log Objects
}

