% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandAvailGrid.r
\name{ExpandAvailGrid}
\alias{ExpandAvailGrid}
\title{Expand grid of availability points}
\usage{
ExpandAvailGrid(polygons = NULL, gridsize = NULL, utm = TRUE)
}
\arguments{
\item{polygons}{SpatialPolygon Object defining the area over which the grid
will be expanded. Typically this will be a shape file of study area with 
different habitat categories defined.}

\item{gridsize}{numeric Size of the grid (in meters)}

\item{utm}{logical Should the grid be in utm (TRUE and default) or 
in longlat (FALSE)}
}
\description{
Expand a grid of availability points over an area defined by a spatial 
polygon (or other spatial object)
}

