% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/format_log_entry.R
\name{format_log_entry}
\alias{format_log_entry}
\title{Default log entry formatter.}
\usage{
format_log_entry(event)
}
\arguments{
\item{event}{The log event to format}
}
\value{
character A string representation of the event for logging.
}
\description{
This function serves as the default formatter for loggr and will
use the format
}
\details{
time - event type - message
}

