% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{indent_log}
\alias{indent_log}
\alias{unindent_log}
\alias{unindent_all_log}
\title{Indenting text}
\usage{
indent_log(indent = "   ")

unindent_log(n = 1)

unindent_all_log()
}
\description{
Indenting text
}
\section{Functions}{
\itemize{
\item \code{indent_log}: Add to existing indent

\item \code{unindent_log}: Remove last indent

\item \code{unindent_all_log}: Remove all indents
}}

