% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_critical}
\alias{log_critical}
\title{Signal a CRITICAL event to log file handlers.}
\usage{
log_critical(message, ...)
}
\arguments{
\item{message}{character: the message to log.}

\item{...}{additional data to log (ignored by default formatter)}
}
\description{
Create and signal a critical condition. And condition handlers listening
for \code{CRITICAL} log events will be notified.
}

