% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_debug}
\alias{log_debug}
\title{Signal a DEBUG event to log file handlers.}
\usage{
log_debug(message, ...)
}
\arguments{
\item{message}{character: the message to log.}

\item{...}{additional data to log (ignored by default formatter)}
}
\description{
Create and signal a debug condition. And condition handlers listening
for \code{DEBUG} log events will be notified.
The message can be wrapped in a one sided formula, in which case
string interpolation will be used.
}

