% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_error}
\alias{log_error}
\title{Signal an error event to log file handlers.}
\usage{
log_error(message, ...)
}
\arguments{
\item{message}{character: the message to log.}

\item{...}{additional data to log (ignored by default formatter)}
}
\description{
Create and signal an error condition. And condition handlers listening
for \code{ERROR} log events will be notified.
}

