% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log_event.R
\name{log_event}
\alias{log_event}
\title{Create a log event}
\usage{
log_event(level, message, ...)
}
\arguments{
\item{level}{character: the event level.}

\item{message}{character: the event message}

\item{...}{further elements to add to the condition.}
}
\value{
A log event (condition)
}
\description{
A log event is a condition and can be signalled as such.
In addition to the classic "simpleMessage", "simpleWarning", and
"simpleError" conditions, loggr allows for levels "INFO", "WARN",
"DEBUG", "ERROR", "CRITICAL".
}

