% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log_file.R
\name{log_file}
\alias{log_file}
\title{Activate a log file}
\usage{
log_file(file_name, ..., .warning = TRUE, .error = TRUE, .message = TRUE,
  .formatter = format_log_entry)
}
\arguments{
\item{file_name}{character: path to the log file.}

\item{...}{list of quoted or unquoted events to log. In none are
provided all log events will be captured.}

\item{.warning}{logical: capture regular warnings (\code{simpleWarning})?}

\item{.error}{logical: capture regular errors (\code{simpleError})?}

\item{.message}{logical: capture regular messages (\code{simpleMessage})?}

\item{.formatter}{logical: the formatting function to use to convert
a log event to its character representation.}
}
\value{
NULL invisibly
}
\description{
Activate a log file
}

