% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log_handler.R
\name{log_handler}
\alias{log_handler}
\title{Condition handler for log events}
\usage{
log_handler(condition)
}
\arguments{
\item{condition}{A condition object.}
}
\description{
The handler will deal with log events for log files setup to receive
notice.
}

