% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_info}
\alias{log_info}
\title{Signal an info event to log file handlers.}
\usage{
log_info(message)
}
\arguments{
\item{message}{character: the message to log.}
}
\description{
Create and signal an info condition. And condition handlers listening
for \code{INFO} log events will be notified.
}

