% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggr.R
\name{log_str}
\alias{log_str}
\title{Log the structure of an object}
\usage{
log_str(x, ...)
}
\arguments{
\item{x}{An R object}

\item{...}{Additional arguments passed to \code{\link{loggr}}}
}
\description{
Log the structure of an object
}

