% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/log_functions.R
\name{log_warn}
\alias{log_warn}
\title{Signal a WARN event to log file handlers.}
\usage{
log_warn(message, ...)
}
\arguments{
\item{message}{character: the message to log.}

\item{...}{additional data to log (ignored by default formatter)}
}
\description{
Create and signal a WARN condition. And condition handlers listening
for \code{WARN} log events will be notified.
}

