% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggr.R
\name{loggr}
\alias{loggr}
\title{Write a log message}
\usage{
loggr(x, ..., .flags = NULL, .level = 0, .file = NULL,
  .timestamp = TRUE)
}
\arguments{
\item{x}{An object.}

\item{...}{Expressions to be logged.}

\item{.flags}{A character vector. These flags are added to the logging statement.
loggr_flags can be used to set this variable globally}

\item{.level}{A integer. Specifies the importance of log. loggr_level
sets the global importance level.Only statements with a level above the global level are
logged}

\item{.file}{A character string or a file connection. If NULL, statements are
printed to standard output connection}

\item{.timestamp}{A boolean. Whether to include a timestamp in the logging statement.}
}
\description{
A pipeable function for logging. The first argument is passed
through the function unchanged, allowing loggr to be used
}
\examples{

iris_filtered = iris \%>\%
 loggr("I see you have, nrow(.), "rows") \%>\%
 filter(Species == "setosa") \%>\%
 loggr("And now you have, nrow(.), "rows")

x = rnorm(100) \%>\%
 loggr(length(.), "samples drawn", .flags=c("flag1", "flag2")) \%>\%

loggr_level(5)

y = x \%>\%
 loggr("This log isn't so important", .level = 1) \%>\%
 max \%>\%
 loggr("But maybe this one is", .level = 10)

}

