% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggr.R
\name{loggr0}
\alias{loggr0}
\title{Write a log message}
\usage{
loggr0(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{loggr}}}
}
\description{
This writes a log message without an input variable
}
\examples{

loggr0("This is a log statement")

}

