% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{loggr_flags}
\alias{loggr_flags}
\title{Get and set global flags for \code{loggr}}
\usage{
loggr_flags(...)
}
\description{
Get or set a vector of flags used for all \code{\link{logger}} calls.
Call with arguments to set flags. Call without arguments to retrive current flags
}
\examples{

loggr_flags("flag1", "flag2")

rnorm(10) \%>\%
 loggr("Message with flags")

loggr_flags()

}
\seealso{
\code{\link{loggr}}, \code{\link{remove_loggr_flags}}
}

