% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{loggr_level}
\alias{loggr_level}
\title{Set verbosity level of \code{loggr}}
\usage{
loggr_level(level = NULL)
}
\arguments{
\item{level}{An integer. The verbosity level}
}
\description{
Sets the global verbosity level of \code{\link{loggr}}. Zero is the most
verbose with increasing numbers indicated less verbosity
}
\seealso{
\code{\link{loggr}}
}

