% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notify_loggr.R
\name{notify_loggr}
\alias{notify_loggr}
\title{Notify loggr about an event.}
\usage{
notify_loggr(..., type = "other")
}
\arguments{
\item{...}{arguments passed to the original event.}

\item{type}{character: either "warning", "error", or "other"}
}
\description{
This is an internal function used by loggr to hook into the relevant
events.
}

