% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_regexp.R
\name{read_regexp}
\alias{read_regexp}
\title{Read a file line by line, only return lines that match the provided regular
expression.}
\usage{
read_regexp(file, regexp, ...)
}
\arguments{
\item{file}{Path to the file}

\item{regexp}{The regular expression}

\item{...}{Remaining arguments will be passed to the readr::read_lines method}
}
\value{
Character vector of lines
}
\description{
Read a file line by line, only return lines that match the provided regular
expression.
}
