% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReclassifyHabitat.r
\name{ReclassifyHabitat}
\alias{ReclassifyHabitat}
\title{Reclassify habitat types}
\usage{
ReclassifyHabitat(Habitat)
}
\arguments{
\item{Habitat}{character The habitat type}
}
\description{
Reclassify raw habitat types into few groups
}

