% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{remove_loggr_flags}
\alias{remove_loggr_flags}
\title{Remove \code{loggr} flags}
\usage{
remove_loggr_flags()
}
\description{
Remove any currently existing flags for \code{\link{loggr}}
}
\examples{

loggr_flags("flag1", "flag2")
loggr_flags()

remove_loggr_flags()
loggr_flags()

}
\seealso{
\code{\link{loggr_flags}}
}

