% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{say}
\alias{say}
\alias{say0}
\alias{sayf}
\alias{say_debug}
\title{Print to screen and log file}
\usage{
say(string, tofile = T)

say0(...)

sayf(fmt, ...)

say_debug(...)
}
\description{
Print to screen and log file
}
\section{Functions}{
\itemize{
\item \code{say}: Print a single string

\item \code{say0}: Print multiple strings with no separation

\item \code{sayf}: Formatted printing to screen and logfile

\item \code{say_debug}: Print strings only to screen
}}

