% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_log_file}
\alias{set_log_file}
\alias{start_log_file}
\alias{append_log_file}
\alias{clear_log_file}
\title{Setting, appending, and clearing log files}
\usage{
set_log_file(filename)

start_log_file(filename)

append_log_file(file)

clear_log_file()
}
\description{
Setting, appending, and clearing log files
}
\section{Functions}{
\itemize{
\item \code{set_log_file}: Set the log file name

\item \code{start_log_file}: Start new log file; overwrite if exists

\item \code{append_log_file}: Start new log file; append if exists

\item \code{clear_log_file}: Delete contents of current log file
}}

