% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set_loggr_hook.R
\name{set_loggr_hook}
\alias{set_loggr_hook}
\title{Set up a loggr hook}
\usage{
set_loggr_hook(which)
}
\arguments{
\item{which}{character: the function to hook up; one of
"message", "warning", "stop", "signalCondition".}
}
\value{
logical indicating success.
}
\description{
This is an internal function used by loggr to set the hooks.
}

