% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/source.R
\name{source}
\alias{source}
\title{Read R Code from a File or a Connection}
\usage{
source(file, local = FALSE, echo = verbose, print.eval = echo,
  verbose = getOption("verbose"), prompt.echo = getOption("prompt"),
  max.deparse.length = 150, chdir = FALSE,
  encoding = getOption("encoding"), continue.echo = getOption("continue"),
  skip.echo = 0, keep.source = getOption("keep.source"))
}
\arguments{
\item{file}{as in \code{base::source}}

\item{local}{as in \code{base::source}}

\item{echo}{as in \code{base::source}}

\item{print.eval}{as in \code{base::source}}

\item{verbose}{as in \code{base::source}}

\item{prompt.echo}{as in \code{base::source}}

\item{max.deparse.length}{as in \code{base::source}}

\item{chdir}{as in \code{base::source}}

\item{encoding}{as in \code{base::source}}

\item{continue.echo}{as in \code{base::source}}

\item{skip.echo}{as in \code{base::source}}

\item{keep.source}{as in \code{base::source}}
}
\description{
This wrapper around \code{\link[base]{source}} allows \code{loggr} to hook
up the log condition handler when files are sourced. Once the handler is
set up, control is handed over to \code{\link[base]{source}}.
}

