% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str}
\alias{str}
\title{Returns structure of an R object}
\usage{
str(x, ...)
}
\arguments{
\item{x}{An R object}

\item{...}{Additional arguments passed to \code{\link{utils::str}}}
}
\description{
This is a version of str that returns the structure of the R object as a character
rather than printing the structure to the terminal
}

