% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/use_logging.R
\name{use_logging}
\alias{use_logging}
\title{Activate logging by loggr}
\usage{
use_logging()
}
\description{
This function hooks up \code{loggr}s handler which is activated by calls
to \code{\link{message}}, \code{\link{warning}}, \code{\link{stop}}, and
\code{\link{signalCondition}}.
}

