% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_log_entry.R
\name{write_log_entry}
\alias{write_log_entry}
\title{Write a log entry to a log file.}
\usage{
write_log_entry(obj, condition)
}
\arguments{
\item{condition}{a condition which can be coersed to a log_event.}

\item{object}{A logger object}
}
\description{
Write a log entry to a log file.
}

