% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReclassifyHabitat.r
\name{ReclassifyHabitat}
\alias{ReclassifyHabitat}
\title{Reclassify habitat types}
\usage{
ReclassifyHabitat(Habitat = NULL, type = NULL)
}
\arguments{
\item{Habitat}{character The habitat type}

\item{type}{character The project - either starling or hare}
}
\description{
Reclassify raw habitat types into few groups
}

