% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write_log_entry.R
\name{write_log_entry}
\alias{write_log_entry}
\title{Write a log entry to a log file.}
\usage{
write_log_entry(file_name, condition, formatter)
}
\arguments{
\item{file_name}{character: the name of the file to write to.}

\item{condition}{a condition which can be coersed to a log_event.}

\item{formatter}{function that returns the character representation of a
log event.}
}
\description{
Write a log entry to a log file.
}

